Renamed some files, pruned some unecessary code, to cut down on console warnings
