+++ src/components/List.js
@@ -1,23 +1,23 @@
-import React, {useState, useEffect} from 'react'
-import axios from 'axios'
+import React, {useState} from 'react'
 import './css/List.css'
 import Card from './Card'
-import Edit_list from './Edit_list'
-import Add_card from './Add_card'
-import { useData, useRefreshData} from '../DataContext'
+import EditList from './EditList'
+import AddCard from './AddCard'
+import { useData } from '../DataContext'
 
 const List = props => {
-    let [list, setList] = useState({...props.list})
     let [listView, setListView] = useState('cards')
+    let [showAddCard, setShowAddCard] = useState(false)
 
-    let thisCards = useData().cards.filter(card => card.list === list.id)
+    let thisCards = useData().cards.filter(card => card.list === props.list.id)
     // let [cards, setCards] = useState([])
 
-    const refreshData = useRefreshData()
-
-
     const showEdit = () => {
-        setListView('Edit_list')
+        setListView('Edit_List')
+    }
+
+    const toggleAddCard = () => {
+        setShowAddCard(!showAddCard)
     }
 
 
@@ -38,12 +38,15 @@ const List = props => {
                             />
                             </>
                 )})}
-                <Add_card list={props.list} />
+                <p className='btn btn-show' onClick={toggleAddCard}>{showAddCard ? 'Close Form' : 'Add Card'}</p>
+                {showAddCard &&
+                    <AddCard list={props.list} />
+                }
                 <button className='btn btn-warning' onClick={showEdit}>Edit List</button>
                 </>
             }
-            {listView === 'Edit_list' &&
-                <Edit_list
+            {listView === 'Edit_List' &&
+                <EditList
                     list={props.list}
                     key={props.list.id}
                     setListView={setListView}
