+++ src/components/ListBox.js
@@ -1,12 +1,12 @@
 import React, {useState} from 'react'
 import {useData} from '../DataContext'
 import List from './List'
-import Add_list from "./Add_list"
+import AddList from "./AddList"
 
 import './css/List.css'
 
 const ListBox = (props) => {
-    const lists = [...useData().lists]
+    const [lists] = useState([...useData().lists])
 
     return(
         <div className="list-box">
@@ -20,7 +20,7 @@ const ListBox = (props) => {
             })}
             <div className='list'>
                 <h4>Add a List</h4>
-                <Add_list />
+                <AddList />
             </div>
         </div>
     )
